Update to the latest design and fix problematic/dead code
