--- hs_auth_proxy/api/__init__.py
@@ -1,1 +1,0 @@
-# HydroShare S3 Auth Proxy API
