+++ hs_micro_auth/Dockerfile
