+++ hs_s3_proxy/.gitignore
