+++ hs_s3_proxy/Dockerfile
