+++ hs_s3_proxy/api/__init__.py
@@ -1,0 +1,1 @@
+# HydroShare S3 Proxy API
