+++ hs_s3_proxy/api/lib/__init__.py
@@ -1,0 +1,1 @@
+# HydroShare S3 Proxy Library
