+++ hs_s3_proxy/requirements.txt
@@ -1,7 +1,5 @@
 fastapi[standard]>=0.113.0,<0.114.0
-psycopg2
 pydantic>=2.7.0,<3.0.0
-sqlalchemy
 boto3>=1.34.0
 httpx>=0.27.0
 python-multipart
