+++ local-dev.yml
@@ -311,7 +311,7 @@ services:
     platform: linux/amd64
     container_name: micro-auth
     build:
-      context: ./hs_auth_minio
+      context: ./hs_micro_auth
       dockerfile: Dockerfile
     environment:
       HS_DATABASE_URL: "postgresql://postgres:postgres@postgis:5432/postgres"
